hand-drawn text mesh
added an addtional mesh to the main mesh location. still needs to be adjusted in scale/position for responsiveness. We might want to add the mesh into the main.js file so it scales with respect to that model vs independently.

I added some doc links and an example codesandbox use

https://github.com/pmndrs/drei#cycleraycast
https://github.com/pmndrs/drei#usecursor

https://codesandbox.io/s/multi-select-edges-forked-xvqxli